imp: bring context types up under src
