+++ ibc-testkit/src/lib.rs
@@ -15,6 +15,7 @@ extern crate alloc;
 #[cfg(feature = "std")]
 extern crate std;
 
+pub mod context;
 pub mod fixtures;
 pub mod hosts;
 pub mod relayer;
